state-module-complete

+++ EMS.Service/Base/BaseService.cs
@@ -1,10 +1,12 @@
 ﻿using AutoMapper;
-using EMS.Entity;
+using EMS.Entity.Base;
 using EMS.Repository.Base;
 using EMS.Service.DTO;
+using EMS.Shared;
 using EMS.Shared.Constant;
 using Microsoft.EntityFrameworkCore;
 using System.Linq.Expressions;
+using EMS.Service.Extension;
 
 namespace EMS.Service.Base
 {
@@ -17,14 +19,24 @@ namespace EMS.Service.Base
             Mapper = mapper;
             Repo = baseRepository; 
         }
-        public async Task AddAsync(D dto)
+        public virtual async Task AddAsync(D dto)
         {
+            if (dto == null) { throw new ArgumentNullException(nameof(dto), "DTO cannot be null"); }
             T entity = ToEntity(dto);
             await Repo.AddAsync(entity);
+        }
+        public virtual async Task UpdateAsync(D dto)
+        {
+            if (dto == null) { throw new ArgumentNullException(nameof(dto), "DTO cannot be null"); }
+            D dtoToUpdate = await GetByIdAsync(dto.Id);
+
+            T entity = ToEntity(dtoToUpdate);
+            await Repo.UpdateAsync(entity);
         }
 
         public async Task DeleteAsync(int Id)
         {
+            if (Id <= 0) { throw new ArgumentNullException(nameof(Id), "Id cannot be null"); }
             await Repo.DeleteAsync(Id);
         }
 
@@ -37,6 +49,8 @@ namespace EMS.Service.Base
 
         public async Task<D> GetByIdAsync(int Id)
         {
+            if (Id <= 0) { throw new ArgumentNullException(nameof(Id), "Id cannot be null"); }
+
             T? entity = await Repo.GetByIdAsync(Id);
             if (entity == null)
             {
@@ -47,31 +61,36 @@ namespace EMS.Service.Base
 
         public async Task<PaginationDTO<D>> GetPageAsync(PaginationDTO<D> paginationDTO)
         {            
-            Expression<Func<D, bool>> expression = paginationDTO.FilterDTO.GetFilter();
-            Expression<Func<T, bool>> where = Map<Expression<Func<D, bool>>, Expression<Func<T, bool>>>(expression);
-            IQueryable<T> query = Repo.GetAll(where);
+            //Expression<Func<D, bool>> expression = paginationDTO.FilterDTO.GetFilter();
+            //Expression<Func<T, bool>> where = Map<Expression<Func<D, bool>>, Expression<Func<T, bool>>>(expression);
+            IQueryable<T> query = Repo.GetAll(null);
 
             paginationDTO.RecordCount = await query.CountAsync();
 
-            query.Skip(paginationDTO.PageSize * (paginationDTO.PageNo - 1));
-            query.Take(paginationDTO.PageSize);
+            if (!string.IsNullOrWhiteSpace(paginationDTO.SortByColumns))
+            {
+                if (paginationDTO.SortBy == EnumSortBy.Descending)
+                {
+                    query = query.OrderByDescending(paginationDTO.SortByColumns);
+                }
+                else
+                {
+                    query = query.OrderBy(paginationDTO.SortByColumns);
+                }
+            }
+
+            query = query.Skip(paginationDTO.PageSize * (paginationDTO.PageNo - 1));
+            query = query.Take(paginationDTO.PageSize);
 
             List<T> records = await query.ToListAsync();
             if (records.Count > 0)
             {
                 paginationDTO.Data = Map<List<T>, List<D>>(records);
-            } else if(paginationDTO.RecordCount > 0 && paginationDTO.PageNo > 1)
-            {
-                paginationDTO.PageNo -= 1;
             }
+
             return paginationDTO;
         }
 
-        public async Task UpdateAsync(D dto)
-        {
-            T entity = ToEntity(dto);
-            await Repo.UpdateAsync(entity);
-        }
 
         #region Protected Methods
         protected virtual D ToDTO(T entity)
